use the property system

+++ tests/test_translations_manager.py
@@ -17,35 +17,29 @@ class TranslationsManagerTestCase(unittest.TestCase):
     """ A classic test class """
 
     def setUp(self):
+        if os.path.exists("./tests/translations_tmp"):
+            shutil.rmtree("./tests/translations_tmp")
         shutil.copytree("./tests/translations", "./tests/translations_tmp")
-        self.translations_manager = TranslationsManager("tests/languages.yml",
-                                                        "tests/translations_tmp/")
+        self.translations_manager = TranslationsManager()
 
     def tearDown(self):
-        shutil.rmtree("./tests/translations_tmp")
+        pass
+        # shutil.rmtree("./tests/translations_tmp")
 
-    def test_constructor(self):
-        """ Test __init__ and get_languages method """
-        self.assertEqual(self.translations_manager.get_languages(),
-                         {"french":"fr", "english":"en", "german":"de"})
 
     def test_create_yaml_files(self):
         """ Test create_yaml_files method """
         shutil.rmtree("./tests/translations_tmp")
         os.mkdir("./tests/translations_tmp")
-        self.translations_manager = TranslationsManager("tests/languages.yml",
-                                                        "tests/translations_tmp")
+        self.translations_manager = TranslationsManager()
 
-        self.translations_manager.create_yaml_files()
+        self.translations_manager._create_yaml_file("./tests/translations_tmp/english.yml")
         dir_files = glob.glob('./tests/translations_tmp/*.yml')
-        dir_files = sorted(dir_files)
-        self.assertEqual(dir_files, ["./tests/translations_tmp/english.yml",
-                                     "./tests/translations_tmp/french.yml",
-                                     "./tests/translations_tmp/german.yml"])
+        self.assertEqual(dir_files, ["./tests/translations_tmp/english.yml"])
 
     def test_load_yaml_files(self):
         """ Test load_yaml_file method """
-        self.translations_manager.load_yaml_files()
+        self.translations_manager.load_yaml_file("english", "./tests/translations_tmp/english.yml")
 
         translateur_en = self.translations_manager.get_translateur("english")
         self.assertEqual(translateur_en.translate("ci_1"), "ci oneéà€")
@@ -54,13 +48,15 @@ class TranslationsManagerTestCase(unittest.TestCase):
         self.assertEqual(translateur_en.translate("ci_4"), "ci four")
         self.assertEqual(translateur_en.translate("ci_5"), "ci five")
 
-    def test_save_in_yaml_files(self):
+    def test_save_yaml_file(self):
         """ Test save_in_yaml_files method """
-        self.translations_manager.load_yaml_files()
+        lang_file = "./tests/translations_tmp/english.yml"
+        self.translations_manager.load_yaml_file("english", lang_file)
         translateur = self.translations_manager.get_translateur('english')
         translateur.add_translation('ci_plop', 'ci_translated')
-        self.translations_manager.save_in_yaml_files()
-        with open("./tests/translations_tmp/english.yml", 'rb') as yml_file:
+        self.translations_manager.save_yaml_file("english", lang_file)
+
+        with open(lang_file, 'rb') as yml_file:
             translations = yaml.safe_load(yml_file)
             if translations == None:
                 translations = {}
@@ -68,17 +64,21 @@ class TranslationsManagerTestCase(unittest.TestCase):
         self.assertEqual(translations['ci_2'], 'ci two')
         self.assertEqual(translations['ci_plop'], 'ci_translated')
 
-    def test_fill_yaml_files(self):
-        """ test fill_yaml_files """
+    def test_fill_yaml_file(self):
+        """ test fill_yaml_file """
         # TODO does not detect rewriting of already translated
 
         ci_manager = CentresOfInterestManager()
         ci_manager.load_xml("tests/ci_two_news.xml")
-        self.translations_manager.fill_yaml_file(ci_manager)
+
+
         translations = {}
         untranslated_token = Translateur.untranslated_token()
-        for lang in self.translations_manager.get_languages():
-            with open("./tests/translations_tmp/"+lang+".yml", 'rb') as yml_file:
+
+        for lang in ['english', 'french']:
+            lang_file = "./tests/translations_tmp/"+lang+".yml"
+            self.translations_manager.fill_yaml_file(lang, lang_file, ci_manager)
+            with open(lang_file, 'rb') as yml_file:
                 translations[lang] = yaml.safe_load(yml_file)
                 if translations[lang] == None:
                     translations[lang] = {}
